requirements.txt
requests
python-dotenv
tomli
pre-commit